requirements.txt
Django==3.2.12
ipython==7.16.1
matplotlib==3.3.4
numpy==1.19.5
opencv_contrib_python==4.5.5.62
Pillow
requests==2.27.1
scikit_image==0.17.2
torch==1.10.2
